docs/implementations/scenario-a-health-check/agent-framework/requirements.txt
# Agent Framework — Health Check Agent dependencies
# Install: pip install -r requirements.txt

agent-framework>=0.1.0
azure-ai-openai>=0.4.0
azure-identity>=1.17.0
python-dotenv>=1.0.0
requests>=2.31.0
